proc: Replace skip_initramfs to androidboot.force_normal_boot
The bootloader always adds skip_initramfs if we boot into Android, and
doesn't add it if we want to boot to recovery.

With retrofitted dynamic partitions we always need the initramfs, but
now androidboot.force_normal_boot is used for booting into Android.

Patch up the cmdline so we can get this behavior.

Issue: FP3-A11#146
Change-Id: I6712cce574e6981577bed6718ad0b3daf57aee2a